Iniciado desenvolvimento da biblioteca para cálculo do algoritmo genético. Realizada melhorias na biblioteca.

+++ src/model/Dados.java
@@ -10,9 +10,9 @@ package model;
  * @author usuario.local
  */
 public class Dados {
-    private int valor;
-    private int peso;
-    private int volume;
+    private double valor;
+    private double peso;
+    private double volume;
 
     public Dados(int valor, int peso, int volume) {
         this.valor = valor;
@@ -24,27 +24,27 @@ public class Dados {
         
     }
     
-    public int getValor() {
+    public double getValor() {
         return valor;
     }
 
-    public void setValor(int valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
 
-    public int getPeso() {
+    public double getPeso() {
         return peso;
     }
 
-    public void setPeso(int peso) {
+    public void setPeso(double peso) {
         this.peso = peso;
     }
 
-    public int getVolume() {
+    public double getVolume() {
         return volume;
     }
 
-    public void setVolume(int volume) {
+    public void setVolume(double volume) {
         this.volume = volume;
     }
